@@ -1,10 +1,8 @@
 export { default as Ecommerce } from './Ecommerce';
-export { default as Kanban } from './Kanban';
+export { default as Login } from './Login';
 export { default as Orders } from './Orders';
 export { default as Employees } from './Employees';
-export { default as Editor } from './Editor';
 export { default as Customers } from './Customers';
-export { default as ColorPicker } from './ColorPicker';
 export { default as Calendar } from './Calendar';
 export { default as Area } from './Charts/Area';
 export { default as Bar } from './Charts/Bar';
@@ -15,3 +13,22 @@ export { default as Pie } from './Charts/Pie';
 export { default as Pyramid } from './Charts/Pyramid';
 export { default as Stacked } from './Charts/Stacked';
 
+// product 
+
+export { default as ProductAdd } from './Product/ProductAdd'
+export { default as Product } from './Product/Product'
+export { default as UpdateProduct } from './Product/UpdateProduct'
+
+// category 
+export { default as Category } from './Category'
+
+// combo
+
+export { default as Combo } from './Combo/Combo'
+export { default as UserRole } from './UserRole/UserRole'
+
+// user
+
+export { default as Users } from './Users/Users'
+export { default as CreateUser } from './Users/CreateUser'
+export { default as UpdateUser } from './Users/UpdateUser'
